Add simple `#gets` based interactor.
This extracts the Readline support into a separate class and keeps only the
generic behavior in the main interactor class, thus enabling the return of
the previous simple `#gets` based interactor.

Guard auto detects if Readline support is suitable on your environment, but
you can also use the new `interactor` DSL method to force a specific
implementation. (Fixes #197 and #200)