Fix terminal status after interrupting the Readline interactor.
The ruby interpreter doesn't restore the status of the terminal if it was interrupted
and leaves the terminal in an unusable state when using thr Readline interactor.
Now when the user interrupts the Readline interactor, the status is restored before stopping.

This commit also makes Guard able to stop nicely when it's interrupted instead of displaying
the whole error trace.