Refactor DSL re-evaluation (Fixes #234)
The issue was fixed by simply turn on the notifier after re-evaluation.

In addition the reevaluate method is uses now a before and after re-
evaluation method, so it can be easily tested (no more pending
examples left).

Also Guard doesn't exit when the Guardfile is empty or invalid, because
it can be re-evaluated after you fixed the Guardfile :P