Rewrite action and scope parsing, add `change` command.
This makes the parsing of scopes and actions form the command
line more flexible. The interactor allows action and groups in any
order, but drops redundant scope or group commands.

    reload rspec

is now the same as

    rspec reload

The interactor command parser returns now any token not matching
a scope or an action, allowing us to get command options. This is
used for the `change` command, to get the name of the spec to run.

    change spec/guard_spec.rb

 will trigger a manual file modification event.