Fix a race condition when stopping Guard
Guard used to delegate blocking the execution to the listener. When the listener
was stopped, execution was resumed on the main thread of Guard. Since there is nothing
more blocking the main thread, the process ended. Any other thread
running was also killed when the process ended.

The Listen turnstile is used now to block execution until Guard is allowed to stop. This allows
any code after stopping the listener to be executed even if it's run from a thread other than
the Guard main thread (like the CLI thread).