Introduce new naming to the README.
We've discussed this at the EuRuKo Hackday and we came to the
conclusion, that we must find a proper name to distinguish
between Guard and a Guard plugin.

Prior to that commit, we often referred to Guard as Guard and
to Guard plugins/implementations as Guards. This doesn't work
in singular, so the confusion when discussing is perfect.

From now, please always strick to the following terms:

Guard (capitalized, singular) => The Guard framework
plugin (lowercase, singular or plural) => A Guard plugin