Update GNTP notifier. (Fixes #231)
With Growl 2.0 being released, the images are shown properly
again, but they most be binary encoded and not a url as
described in the GNTP spec. This can be achieved by using
`127.0.0.1` instead of `localhost`, and ruby_gntp will send
the image as binary.

In addition, on GNTP registration Guard send its application
icon, so we have the nice Guard logo in the Growl preferences.