Reverse UI and command execution debug stubs.
This inverses the testing approach for Guard::UI
and Guard.debug_command_execution. Now
the spec helper stubs out some methods, so that
the specs doesn't have to care about the UI logs
flooding the spec output or that a multiple command
execution debugging causes a stack overflow, thus
before testing the stubbed methods, we must unstub it.