Remove 1.9 Syntax from the README.
The current README only makes use of the Ruby 1.8 syntax
and we should continue to do so, because the syntax works
also fine on 1.9 and it doesn't confuse non Ruby people why
there are two different syntax available.

In about 6 month we're going to drop Ruby 1.8.7 support anyway,
and we'll convert everything then.