Restructure README.
This moves some information to the wiki and changes some headings
to `h4`, so they won't appear on the guardgem.org menu.

The idea is to make the menu on the Guard homepage more slim and
focus on the important parts. Advances and rarely used stuff needs to
go to the wiki.