Change quiet handling for Tmux <= 1.6
Prior to Guard 2.2 the Tmux notifier set the `quiet` option to `on` when
notifications are turned on and to `off` when turned off. This
interferes with custom configuration and other tools, so with Guard 2.2
the output of the commands is set to quiet on a command basis, which
solves the issue.

But for people running the old Tmux version 1.6, this breaks the Tmux
notifier since since the `set-option` command does support the `-q`
flag only since version (See #512).

This commit reverts some testing to work around the backward
compatibility issue and commits to the new approach with one exception:
the title notification is not quiet on Tmux <= 1.6, since it's the only
part which relies on `set-option`. This implies that Tmux 1.6 users
with the title notification set to true are seeing the Tmux command
output. Now it's up to the user to configure that Tmux is quiet by
adding

```
set-option -s quiet
```

to their Tmux config.