Integrate the reload functionality
This uses the advertised solution but it integrates it into the gem.
`bin/guard` is now the wrapper and it and it the original guard is now
`bin/_guard-core`.

All parameters to the wrapper are passed down to `_guard-core` --> all
tests are green.

I think this is much better from a user's perspective.